Billind_and_inventory_software/requirements.txt
mysql-connector-python==8.3.0
reportlab==4.1.0
